+++ lib/motion-prime.rb
@@ -5,6 +5,7 @@ require 'sugarcube-common'
 require 'bubble-wrap'
 require 'bubble-wrap/reactor'
 require 'rm-digest'
+require 'afmotion'
 require File.expand_path('../../motion-prime/env.rb', __FILE__)
 require File.expand_path('../../motion-prime/prime.rb', __FILE__)