+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+=== 0.7.0
+* Added Model.find(1) syntax support where 1 is :id attribute.
+* Migrate to AFMotion.
+
 === 0.6.0
 * Fix executable.
 * Stable release.